# openapi.yaml
openapi: 3.1.0
info:
  title: E-Signing Service API
  description: |
    Backend service pro elektronické podepisování dokumentů.

    ## Autentifikace

    API používá tři typy autentifikace:

    1. **Google ID Token** - Pro authenticated endpointy (`/v1/me`, `/v1/documents/*`)
       - Header: `Authorization: Bearer {google_id_token}`
       - Header: `X-Workspace-ID: {workspace_uuid}`

    2. **Admin Secret + User ID** - Pro Edge Function volání
       - Header: `X-Admin-Secret: {admin_secret}`
       - Header: `X-User-ID: {user_supabase_uuid}`
       - Header: `X-User-Email: {user_email}` (optional)
       - Header: `X-User-Name: {user_name}` (optional)
       - Header: `X-Workspace-ID: {workspace_uuid}`

    3. **Magic Link Token** - Pro signing session endpointy (`/v1/signing/*`)
       - Token je součástí URL path parametru
       - Veřejné endpointy, nevyžadují autentifikaci

    ## Rate Limiting

    - `/v1/verify/*` endpointy: 10 požadavků/minutu per IP
    - `/v1/signing/sessions/*/otp/send`: 5 OTP/hodinu per session, 60s cooldown
    - `/v1/signing/sessions/*/otp/verify`: 5 pokusů, pak 15min lock
  version: 2.0.0
  contact:
    name: E-Signing Service Support
servers:
  - url: https://e-signing-service-526681856440.europe-west1.run.app
    description: Production server
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Google ID Token
      description: Google ID Token získaný přes OAuth2 flow

    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin API secret pro Edge Function autentifikaci

    UserID:
      type: apiKey
      in: header
      name: X-User-ID
      description: User Supabase UUID (required s X-Admin-Secret)

    WorkspaceID:
      type: apiKey
      in: header
      name: X-Workspace-ID
      description: Workspace UUID pro multi-tenant přístup
...
  schemas:
...
    AuthGoogleRequest:
      type: object
      required:
        - code
        - redirect_uri
      properties:
        code:
          type: string
          description: OAuth2 authorization code z Google
        redirect_uri:
          type: string
          description: Redirect URI použitý při autorizaci
...
    AuthenticatedUser:
      type: object
      properties:
        user_id:
          type: string
          description: Google 'sub' identifikátor
        internal_user_id:
          type: string
          format: uuid
          description: Interní Supabase UUID
        workspace_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        picture:
          type: string
          format: uri
        role:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          default: true
        code:
          type: string
          description: Kód chyby (VALIDATION_ERROR, NOT_FOUND, RATE_LIMIT_EXCEEDED, ...)
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
          format: uuid
...
paths:
...
  /v1/auth/google:
    post:
      tags: [auth]
      summary: Google OAuth2 token exchange
      description: Vyměňuje OAuth2 authorization code za Google ID Token
      operationId: authGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGoogleRequest'
      responses:
        '200':
          description: Token exchange successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_token:
                    type: string
                    description: Google ID Token pro další API volání
                  user:
                    $ref: '#/components/schemas/AuthenticatedUser'
        '400':
          description: Invalid authorization code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # Users
  # ============================================================================
...